requirements.txt
argon2==0.1.10
psycopg2==2.9.5
pycryptodome==3.17
python-dotenv==0.21.1
requests==2.28.2
